# backend/app/admin/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: GoWind Admin API
    description: GoWind Admin API
    contact:
        name: tx7do
        url: https://github.com/tx7do/go-wind-admin
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/go-wind-admin/blob/master/LICENSE
    version: "1.0"
paths:
...
    /admin/v1/tasks:
        get:
            tags:
                - TaskService
            description: 查询调度任务列表
            operationId: TaskService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: query
                  in: query
                  description: JSON字符串过滤条件，基础语法：{"field1":"val1", "field2___icontains":"val2"}，具体请参见：https://github.com/tx7do/go-crud/tree/main/pagination/filter/README.md
                  schema:
                    type: string
                - name: filter
                  in: query
                  description: Google AIP规范字符串过滤条件
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: string
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListTaskResponse'
        post:
            tags:
                - TaskService
            description: 创建调度任务
            operationId: TaskService_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateTaskRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/tasks/type-name/{typeName}:
        get:
            tags:
                - TaskService
            description: 查询调度任务详情
            operationId: TaskService_Get
            parameters:
                - name: typeName
                  in: path
                  required: true
                  schema:
                    type: string
                - name: id
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Task'
    /admin/v1/tasks/{id}:
        get:
            tags:
                - TaskService
            description: 查询调度任务详情
            operationId: TaskService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: typeName
                  in: query
                  schema:
                    type: string
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Task'
        put:
            tags:
                - TaskService
            description: 更新调度任务
            operationId: TaskService_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateTaskRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
        delete:
            tags:
                - TaskService
            description: 删除调度任务
            operationId: TaskService_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/tasks:control:
        post:
            tags:
                - TaskService
            description: 控制调度任务
            operationId: TaskService_ControlTask
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ControlTaskRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/tasks:restart:
        post:
            tags:
                - TaskService
            description: 重启所有的调度任务
            operationId: TaskService_RestartAllTask
            requestBody:
                content:
                    application/json: {}
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RestartAllTaskResponse'
    /admin/v1/tasks:start:
        post:
            tags:
                - TaskService
            description: 启动所有的调度任务
            operationId: TaskService_StartAllTask
            requestBody:
                content:
                    application/json: {}
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/tasks:stop:
        post:
            tags:
                - TaskService
            description: 停止所有的调度任务
            operationId: TaskService_StopAllTask
            requestBody:
                content:
                    application/json: {}
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
...
        ControlTaskRequest:
            type: object
            properties:
                controlType:
                    enum:
                        - Start
                        - Stop
                        - Restart
                    type: string
                    description: 控制类型
                    format: enum
                typeName:
                    type: string
                    description: 任务执行类型名，例如 "send_email"、"generate_report" 等，用于区分不同类型的任务
            description: 控制调度任务 - 请求
...
        CreateTaskRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/Task'
            description: 创建调度任务 - 请求
...
        ListTaskResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Task'
                total:
                    type: string
            description: 查询调度任务列表 - 回应
...
        RestartAllTaskResponse:
            type: object
            properties:
                count:
                    type: integer
                    format: int32
            description: 重启调度任务 - 回应
...
        Task:
            type: object
            properties:
                id:
                    type: integer
                    description: 任务ID
                    format: uint32
                type:
                    enum:
                        - PERIODIC
                        - DELAY
                        - WAIT_RESULT
                    type: string
                    description: 任务类型
                    format: enum
                typeName:
                    type: string
                    description: 任务执行类型名，例如 "send_email"、"generate_report" 等，用于区分不同类型的任务
                taskPayload:
                    type: string
                    description: 任务数据，以 JSON 格式存储，方便存储不同类型和数量的参数
                cronSpec:
                    type: string
                    description: cron表达式，用于定义任务的调度时间
                taskOptions:
                    $ref: '#/components/schemas/TaskOption'
                enable:
                    type: boolean
                    description: 启用/禁用任务
                remark:
                    type: string
                    description: 备注
                tenantId:
                    type: integer
                    description: 租户ID，0代表系统全局角色
                    format: uint32
                createdBy:
                    type: integer
                    description: 创建者ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 调度任务
        TaskOption:
            type: object
            properties:
                maxRetry:
                    type: integer
                    description: 任务最多可以重试的次数
                    format: uint32
                timeout:
                    pattern: ^-?(?:0|[1-9][0-9]{0,11})(?:\.[0-9]{1,9})?s$
                    type: string
                    description: 任务超时时间
                deadline:
                    type: string
                    description: 任务截止时间
                    format: date-time
                processIn:
                    pattern: ^-?(?:0|[1-9][0-9]{0,11})(?:\.[0-9]{1,9})?s$
                    type: string
                    description: 任务延迟处理时间
                processAt:
                    type: string
                    description: 任务执行时间点
                    format: date-time
                uniqueTTL:
                    pattern: ^-?(?:0|[1-9][0-9]{0,11})(?:\.[0-9]{1,9})?s$
                    type: string
                    description: 任务唯一锁定时间
                retention:
                    pattern: ^-?(?:0|[1-9][0-9]{0,11})(?:\.[0-9]{1,9})?s$
                    type: string
                    description: 任务结果保留时间
                group:
                    type: string
                    description: 任务分组
                taskID:
                    type: string
                    description: 任务唯一标识ID
            description: 任务选项
...
        UpdateTaskRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                data:
                    $ref: '#/components/schemas/Task'
                updateMask:
                    example: id,realname,username
                    type: string
                    description: 要更新的字段列表
                    format: field-mask
                allowMissing:
                    type: boolean
                    description: 如果设置为true的时候，资源不存在则会新增(插入)，并且在这种情况下`updateMask`字段将会被忽略。
            description: 更新调度任务 - 请求
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh-token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...